A login form in JSP
A login form created using JSP to show details of companies that visited the campus in the previous 3 years along with the number of candidates they hired

The LOGIN Form checks if the username matches the entries from the database and the password is the current date in DDMMYYYY format. On successful login, You can select a batch and a table will display the list of companies and the number of candidates each of them hired.